+++ webApi/Models/PropertyDetails.cs
@@ -7,9 +7,9 @@ public class PropertyDetails
     public int PropertyDetailsId { get; set; }
     public required string PropertyName { get; set; }
     public required string Address { get; set; }
-    public decimal LandSizeInSquareMeters { get; set; }
-    public decimal ConstructionSizeInSquareMeters { get; set; }
-    public decimal NumberOfRooms { get; set; }
+    public double LandSizeInSquareMeters { get; set; }
+    public double ConstructionSizeInSquareMeters { get; set; }
+    public double NumberOfRooms { get; set; }
     public required string Description { get; set; }
     public string Photo { get; set; } = "https://placehold.co/300x200/d1d1d1/dedede?text=img";
     public int PropertyId { get; set; }
